+++ src/types.ts
@@ -26,6 +26,7 @@ export type DataType = 'string' | 'number' | 'boolean' | 'array' | 'object' | 'b
 export interface ValidatorConfig {
 	quotes?: 'none' | 'single-quotes' | 'double-quotes' | 'backtick';
 	strict?: boolean;
+	arrayIndexingCheck?: boolean;
 }
 
 export type ConfigType = `field:${string}` | `error:${string}`;